proto: use XXX_unrecognized exclusively for unknown fields
The protobuf data model makes no distinction between unknown fields
that are within the extension field ranges or not. Now that we eagerly
unmarshal extensions, there is even less need for storing unknown
fields in the extension map. Instead, use the XXX_unrecognized field
exclusively for this purpose.

To support this logic, we fork the v2 internal/encoding/wire package.
This is a temporary measure since the v1 code will be completely
re-written in terms of the v2 API in the near future.

Change-Id: I3dadd04ec2314e6d245d46f6329383bb9e0d00f7
Reviewed-on: https://go-review.googlesource.com/c/protobuf/+/175580
Reviewed-by: Damien Neil <dneil@google.com>