protoc-gen-go: add stability warning to XxxServer and XxxClient
Due to the lack of stability guarantees provided for service Server and
Client interfaces, as hightlighted in grpc/grpc-go#2318 and
grpc/grpc-go#3024, this change adds a godoc line that both warns users
and suggests the canonical workaround.

The structure of the godoc text is styled after an open proposal,
golang/go#34409, thus may need to be changed, but attempts to play well
with the tooling ecosystem, as was done for deprecated with
golang/go#10909.