generate json tags for oneof fields
While protoc-gen-go generates JSON tags for most message structure members, it
does not do so when the member represent a oneof construct or when the
structure represent an element of a oneof member.

This means that programs using encoding/json instead of jsonpb will end up
with incoherent names which do not match the name of the Protobuf fields. They
will also be impredictable due to alternate names used to avoid collisions or
syntax errors.

Adding these tags makes the generated code coherent.