proto: add text tests
These tests were removed during an era when we thought that
the v1 text implementation would just wrap the v2 implementation.
That didn't happen since too much code depended on the exact behavior
of the v1 implementation (exact output, bugs, error strings, etc).

Add the tests back in.
This is a cleaned up version of the tests from v1.3.3.

Change-Id: I0419af2092c4ed5703636366844ea61e910c5908
Reviewed-on: https://go-review.googlesource.com/c/protobuf/+/220351
Reviewed-by: Damien Neil <dneil@google.com>