add support for explicitly marking tag ids reserved
By using the new protobuf3.Reserved type and listing the reserved
IDs in the protobuf tag value. This is better than a comment for
two reasons: 1) we verify when parsing the tags that reserved
IDs are not used elsewhere, and 2) when generating the protobuf
message we add a `reserved ...;`  keyword, passing the reserved
IDs on to later tools.