improve the encoding of structs where all the elements
encoded to nothing at all. we used to do like the official protobuf
package and always encode these as {tag, len=0}. But that is not
necessary if the struct is not a member of a repeated element
(slice of struct, slice of pointer-to-struct, for example).

Even a map with empty key or value can be elided this way. The
official decoder supports it.